add user relation to todo routes

+++ app/api/todos/[id]/route.ts
@@ -7,11 +7,25 @@ export async function GET(
   req: NextApiRequest,
   { params }: { params: { id: string } }
 ) {
-  const { isAuthenticated } = await auth(req);
-  if (isAuthenticated) {
+  const { isAuthenticated, token } = await auth(req);
+
+  if (isAuthenticated && token?.email) {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: token.email,
+      },
+    });
+
+    if (!user) {
+      return new Response(JSON.stringify({ error: "User not found" }), {
+        status: 404,
+      });
+    }
+
     const todo = await prisma.todo.findUnique({
       where: {
         id: parseInt(params.id, 10),
+        userId: user.id,
       },
     });
 
@@ -34,8 +48,23 @@ export async function PUT(
   req: Request,
   { params }: { params: { id: string } }
 ) {
-  const { isAuthenticated } = await auth(req as unknown as NextApiRequest);
-  if (isAuthenticated) {
+  const { isAuthenticated, token } = await auth(
+    req as unknown as NextApiRequest
+  );
+
+  if (isAuthenticated && token?.email) {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: token.email,
+      },
+    });
+
+    if (!user) {
+      return new Response(JSON.stringify({ error: "User not found" }), {
+        status: 404,
+      });
+    }
+
     const { title, completed } = await req.json();
 
     if (
@@ -49,19 +78,27 @@ export async function PUT(
       });
     }
 
-    const updatedTodo = await prisma.todo.update({
-      where: {
-        id: parseInt(params.id, 10),
-      },
-      data: {
-        title,
-        completed,
-      },
-    });
+    try {
+      const updatedTodo = await prisma.todo.update({
+        where: {
+          id: parseInt(params.id, 10),
+          userId: user.id,
+        },
+        data: {
+          title,
+          completed,
+        },
+      });
 
-    return new Response(JSON.stringify(updatedTodo), {
-      status: 200,
-    });
+      return new Response(JSON.stringify(updatedTodo), {
+        status: 200,
+      });
+    } catch (error) {
+      // If the todo item is not found, Prisma will throw an error
+      return new Response(JSON.stringify({ error: "Todo not found" }), {
+        status: 404, // Not Found
+      });
+    }
   } else {
     return new Response(JSON.stringify({ error: "Unauthorized" }), {
       status: 401,
@@ -74,13 +111,33 @@ export async function DELETE(
   req: NextApiRequest,
   { params }: { params: { id: string } }
 ) {
-  const { isAuthenticated } = await auth(req);
-  if (isAuthenticated) {
-    await prisma.todo.delete({
-      where: { id: parseInt(params.id, 10) },
+  const { isAuthenticated, token } = await auth(req);
+
+  if (isAuthenticated && token?.email) {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: token.email,
+      },
     });
 
-    return new Response(null, { status: 204 });
+    if (!user) {
+      return new Response(JSON.stringify({ error: "User not found" }), {
+        status: 404,
+      });
+    }
+
+    try {
+      await prisma.todo.delete({
+        where: { id: parseInt(params.id, 10), userId: user.id },
+      });
+
+      return new Response(null, { status: 204 }); // No Content
+    } catch (error) {
+      // If the todo item is not found, Prisma will throw an error
+      return new Response(JSON.stringify({ error: "Todo not found" }), {
+        status: 404, // Not Found
+      });
+    }
   } else {
     return new Response(JSON.stringify({ error: "Unauthorized" }), {
       status: 401,
